refactor: turned nav with functions into component
resolves #359
references #350
