+++ src/routes/story/[id]/+page.svelte
@@ -1,10 +1,7 @@
 <script>
-  import { Back, VisualsSVG, CloudsSVG, DarkModeSVG } from '$lib/index';
-  import { onMount } from 'svelte';
-  import { enhance } from '$app/forms';
-
+  import { CloudsSVG, StoryNavActions } from '$lib/index';
   export let data;
-  const { story, audio, theme, showVisuals: showVisualsFromServer } = data;
+  const { story, audio} = data;
   const audioSrc = story.audios?.[0]?.file || '';
 
   let transcriptLines = audio?.transcript ? parseVTT(audio.transcript) : [];
@@ -13,43 +10,11 @@
   let currentLineIndex = -1;
   let transcriptRefs = [];
 
-  let showVisuals = showVisualsFromServer;
-  let darkMode = theme === 'dark';
+  let showVisuals = data.showVisuals;
+  let darkMode = data.theme === 'dark';
   let jsEnabled = false;
 
-  onMount(() => {
-    jsEnabled = true;
-    const storedVisuals = localStorage.getItem('showVisuals');
-    const storedTheme = localStorage.getItem('theme');
-
-    if (storedVisuals !== null) showVisuals = storedVisuals === 'true';
-    if (storedTheme) darkMode = storedTheme === 'dark';
-  });
-
-  function toggleVisuals() {
-    if (!document.startViewTransition) {
-      showVisuals = !showVisuals;
-      localStorage.setItem('showVisuals', showVisuals);
-      return;
-    }
-    document.startViewTransition(() => {
-      showVisuals = !showVisuals;
-      localStorage.setItem('showVisuals', showVisuals);
-    });
-  }
-
-  function toggleTheme() {
-    if (!document.startViewTransition) {
-      darkMode = !darkMode;
-      localStorage.setItem('theme', darkMode ? 'dark' : 'light');
-      return;
-    }
-    document.startViewTransition(() => {
-      darkMode = !darkMode;
-      localStorage.setItem('theme', darkMode ? 'dark' : 'light');
-    });
-  }
-
+ 
   function parseVTT(vtt) {
     const lines = vtt.split('\n');
     const result = [];
@@ -104,48 +69,12 @@
       el.scrollIntoView({ behavior: 'smooth', block: 'center' });
     }
   }
+
 </script>
 
 <main class:light-mode={!darkMode} style="view-transition-name:main-bg;">
-  <header>
-    <a href="/lessons" aria-label="Go back">
-      <Back color={darkMode ? 'white' : 'black'} />
-    </a>
-
-    <div class="actions">
-      <form method="POST" action="?/toggleTheme" use:enhance on:submit|preventDefault={toggleTheme}>
-        <button aria-label="Toggle theme" type="submit">
-          <DarkModeSVG {darkMode} />
-        </button>
-      </form>
-
-      <form method="POST" action="?/toggleVisuals" use:enhance on:submit|preventDefault={toggleVisuals}>
-        <button aria-label="Toggle visuals" type="submit">
-          {#if showVisuals}
-            <VisualsSVG mode="off" />
-          {:else}
-            <VisualsSVG mode="on" />
-          {/if}
-        </button>
-      </form>
-    </div>
-  </header>
-
-  {#if showVisuals}
-    <section class="visuals">
-      <picture class="story-image flex-items">
-        <source srcset="{story.image}?width=320&format=avif" type="image/avif">
-        <source srcset="{story.image}?width=320&format=webp" type="image/webp">
-        <source srcset="{story.image}?width=320" type="image/jpeg">
-        <img
-          src="{story.image}?width=320"
-          alt="{story.summary} cover image"
-          height="270"
-          width="320"
-        />
-      </picture>
-    </section>
-  {/if}
+ 
+	<StoryNavActions bind:showVisuals bind:darkMode {story}/>
 
   <section class="transcript">
     <h2>{story.title}</h2>
@@ -180,13 +109,6 @@
     {/if}
   </section>
 
-  {#if darkMode}
-    <CloudsSVG style="view-transition-name:clouds" color="dark" />
-    <div class="moon" style="view-transition-name:clouds"></div>
-  {:else}
-    <CloudsSVG style="view-transition-name:clouds" color="light" />
-    <div class="sun" style="view-transition-name:clouds"></div>
-  {/if}
 </main>
 
 <style>
@@ -206,69 +128,9 @@ main.light-mode {
   background: var(--bg-image-blue);
 }
 
-.sun,
-.moon {
-  position: absolute;
-  top: -2em;
-  left: -1em;
-  width: 10em;
-  height: 10em;
-  border-radius: 50%;
-  box-shadow: 0 0 30px #ffcc00;
-  animation: rise 1s ease-out forwards;
-  z-index: -1;
-}
-
-.sun {
-  background: radial-gradient(circle, #ffe066, #ffcc00);
-}
-
-.moon {
-  background: radial-gradient(circle, #d3d3d3, #838383);
-  box-shadow: 0 0 30px #d6d6d6;
-}
-
-header, .visuals, .transcript, .player {
+.transcript, .player {
   padding: 1em;
   max-width: 31.25em;
-}
-
-header {
-  display: flex;
-  width: 100%;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-}
-
-header a {
-  font-size: 1.5em;
-  color: white;
-  text-decoration: none;
-}
-
-.actions {
-  display: flex;
-}
-
-.actions button {
-  background: none;
-  border: none;
-  color: white;
-  font-size: 1.2em;
-  margin-left: 0.5em;
-  cursor: pointer;
-}
-
-.visuals {
-  margin: auto;
-}
-
-.visuals img {
-  max-width: 20em;
-  max-height: 20em;
-  border-radius: 1em;
-  transition: 1s;
 }
 
 .transcript {
@@ -334,17 +196,6 @@ main.light-mode .transcript p {
 
 .player audio {
   width: 100%;
-}
-
-@keyframes rise {
-  from {
-    opacity: 0;
-    transform: translateY(2em);
-  }
-  to {
-    opacity: 1;
-    transform: translateY(0);
-  }
 }
 
 ::view-transition-old(main-bg),
